pyproject.toml
[project]
name = "sample-aws-strands-nova-voice-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=3.4.3",
    "boto3>=1.28.0",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
    "strands-agents[bidi]>=0.1.0",
    "strands-agents-tools>=0.1.0",
    "websockets==10.4",
]
